% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_score_matches.R
\name{get_and_evaluate_matches}
\alias{get_and_evaluate_matches}
\title{Get and evaluate matches using heuristic and LP methods}
\usage{
get_and_evaluate_matches(
  mentee_data,
  mentor_data,
  scoring_criteria,
  weights = NULL,
  matching_criteria,
  valIfNA = 0,
  compare_original = FALSE,
  original_matches_clean = NULL,
  original_name = "original",
  verbose = TRUE,
  saveAllData = FALSE,
  saveTimes = FALSE
)
}
\arguments{
\item{scoring_criteria}{df of criteria used for scoring matches}

\item{weights}{df weights given to each row in scoring criteria (optional)}

\item{matching_criteria}{df of criteria used in heuristic}

\item{valIfNA}{default = 0}

\item{compare_original}{boolean. If TRUE compares heuristic and LP to original data}

\item{verbose}{boolean}

\item{m_ee}{df of mentee data}

\item{m_or}{df of mentor data}

\item{original_matches_raw}{optional df containing matches}
}
\value{
df showing performance of each method
}
\description{
Get and evaluate matches using heuristic and LP methods
}
