% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_score_matches.R
\name{get_best_matches}
\alias{get_best_matches}
\title{get_best_match
after running algorithm on multiple versions of rule and LP}
\usage{
get_best_matches(results, criteria = "mean_score")
}
\arguments{
\item{criteria}{}
}
\description{
get_best_match
after running algorithm on multiple versions of rule and LP
}
