% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_score_matches.R
\name{get_LP_based_matches}
\alias{get_LP_based_matches}
\title{LP model for matches}
\usage{
get_LP_based_matches(
  scores,
  m_ee,
  m_or,
  num_mentees_per_mentor = 1,
  num_mentors_per_mentee = 1,
  valIfNA = 0,
  score_column_name = "total_score"
)
}
\arguments{
\item{scores}{dataframe with mentee Name, mentor name and total_score column}

\item{m_ee}{mentee data}

\item{m_or}{mentors data}
}
\description{
LP model for matches
}
