% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_score_matches.R
\name{get_rule_based_matches_all}
\alias{get_rule_based_matches_all}
\title{Title}
\usage{
get_rule_based_matches_all(m_ee_in, m_or_in, matching_criteria_in)
}
\arguments{
\item{m_ee_in}{mentee dataframe. must contain `mentee_name` column as unique identifier of each mentee}

\item{m_or_in}{mentor dataframe. must contain `mentor_name` column as unique identifier of each mentor}

\item{mentee_criteria_list_in}{a list containing vectors of column names to match on}

\item{mentor_criteria_list_in}{a list containing vectors of column names to match on}

\item{mentee_filter_list_in}{a list containing NULL or a string containing a filtering condition for mentee data, e.g. "gender_pref == 'no_pref' "}

\item{mentor_filter_list_in}{a list containing NULL or a string containing a filtering condition for mentor data}
}
\value{
either a tibble of all mentees with their matches, or if demo =TRUE, a list of  matches + other things.
}
\description{
Title
}
\details{
matches mentees and mentors based on column names provided in mentee_criteria_in and mentor_criteria_in.
}
