% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_score_matches.R
\name{get_unmatched_mentees}
\alias{get_unmatched_mentees}
\title{Get unmatched mentees}
\usage{
get_unmatched_mentees(matches_tb)
}
\arguments{
\item{matches_tb}{a tibble of matches}
}
\value{
a tibble of unmatched mentees
}
\description{
Get unmatched mentees
}
