% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_score_matches.R
\name{get_unmatched_mentors}
\alias{get_unmatched_mentors}
\title{Get unmatched mentors}
\usage{
get_unmatched_mentors(matches_tb, m_or_tb)
}
\arguments{
\item{matches_tb}{a tibble of matches}

\item{m_or_tb}{a tibble of mentors}
}
\value{
a tibble of unmatched mentors
}
\description{
Get unmatched mentors
}
